fix: Applying TAX to Multiple Product
Changes: Use separate class for tax and product and removed ProductUtilities class
which holds everything inside it, and i have introduced a new class called Report class with tasks related to the report
printing.
also includes a new class called StringFormatter to deal with formatting values in the required and right way.